Fix error thrown when invalid city name is given (#147)
Basically, if a city is invalid, `WeatherRemoteDataSource#getWeather` would return
`InvalidCityName`. The old code would then memoize the weather if it's
not `Left`, but if it is, then an error would be thrown (oversight from
my side). The issue is also present in `ForecastsRepo`, so the change is
made there too.